add app controller

+++ src/displayapp/screens/Tile.cpp
@@ -51,7 +51,7 @@ Tile::Tile(uint8_t screenID,
     if (i == 3) {
       btnmMap[btIndex++] = "\n";
     }
-    if (applications[i].application == Apps::None) {
+    if (applications[i].application == static_cast<uint8_t>(Apps::None)) {
       btnmMap[btIndex] = " ";
     } else {
       btnmMap[btIndex] = applications[i].icon;
@@ -76,7 +76,7 @@ Tile::Tile(uint8_t screenID,
 
   for (uint8_t i = 0; i < 6; i++) {
     lv_btnmatrix_set_btn_ctrl(btnm1, i, LV_BTNMATRIX_CTRL_CLICK_TRIG);
-    if (applications[i].application == Apps::None) {
+    if (applications[i].application == static_cast<uint8_t>(Apps::None)) {
       lv_btnmatrix_set_btn_ctrl(btnm1, i, LV_BTNMATRIX_CTRL_DISABLED);
     }
   }
